Add files via upload
Comparison of LSTM and GRU E-commerce for classification of Clothing reviews data
Customer reviews have been commonly recognized as valuable sources for marketing intelligence and sentiment analysis. Sentiment analysis seeks to build a system for analyzing and evaluating customer reviews reflected on websites, blogs, Twitter, or Instagram. Communication has played a key role in boosting social relationships since historical times.
In today's web-based world, virtually everyone is reading online reviews. In fact, 91% of people read them and 84% trust them as much as they would a personal recommendation. The effects of reviews are measurable, too. The average customer is willing to spend 31% more on a retailer that has excellent reviews. Negative reviews can carry as much weight as positive ones. One study found that 82% of those who read online reviews specifically seek out negative reviews. That may sound alarming - this stat only emphasizes that negative reviews aren't going unnoticed - but there are some benefits: Research indicates that users spend five times as long on sites when interacting with negative reviews, with an 85% increase in conversion rate.
Women's E-commerce clothing reviews dataset have been used. This dataset includes 23486 rows and 10 feature variables(link)
Instead of manually tagging the review texts, the sentiment analyzer of NLTK was used to automate the process. The mentioned manual, intuitive tagging had the flaw of not taking into account some neutral sentiments. Hence, the use of sentiment analyzer by NLTK. The GloVe word embeddings were used to map the words in review texts to the vector space.
GloVe EmbeddingEDA:
Distribution of Age vs Count of reviewsIt is found out that women between the age group 30 and 39 become financially independent during this phase and they shop a lot.
After applying all the NLP, the data is ready to get trained with the help of deep learning models. In this project, Bi-directional Long Short-Term Memory (LSTM) and Gated Recurrent Unit (GRU)
are used for training the data. An LSTM model for sentiment classification is generated and complied. Next the training phase begins and the model learns how to classify the reviews with experience by the end of the training phase. We can see how the performance progresses with the experience in the epochs. The number of hidden layers used in the LSTM are 7 and the activation functions used are relu and sigmoid. The loss function is binary cross entropy since the output is either 0 or 1. The optimizer is Adam. The learning rate is 0.0001 and the
metrics used is AUC. The number of epochs is 60 and the batch size is 252.The performance increases epoch by epoch indicating that the model learns on experience. We can see the accuracy and loss of the learning phase. After this, the Gated Recurrent Unit is applied in which the same thing happens as mentioned above in the LSTM. For the GRU, just hidden layers are used. The activation function used is relu in the first layer and the sigmoid in the last layer. The loss function used is binary cross entropy and the optimizer used is Adam and learning rate is 0.006. The number of epochs is 10.
The training accuracy of LSTM model is 94% and the we can see that loss has been gradually decreasing for training and also validation dataset .The accuracy of the GRU model is 99% at the 30th epoch and it went down to 98% at the 10th epoch.
It is evident that GRU uses less training parameter and uses less memory and it executes faster than LSTM. To validate the model manually, I have created a test dataset with 10 reviews to check the model performance of each of the model. GRU predicted almost all the reviews accurately while LSTM failed in 3 cases